Switch svg loader for easier use.
