+++ context/app/static/js/App.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
 import Details from './components/Details'
 import NoticeAlert from "./components/NoticeAlert";
