+++ context/app/static/js/components/VisTabs.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Box from '@material-ui/core/Box';
-import makeStyles from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
 import ProvGraph from './ProvGraph';
 import Tab from '@material-ui/core/Tab';
