+++ package.json
@@ -13,11 +13,11 @@
     "@babel/core": "^7.8.4",
     "@babel/preset-env": "^7.8.4",
     "@babel/preset-react": "^7.8.3",
+    "@svgr/webpack": "^5.2.0",
     "babel-loader": "^8.0.6",
     "css-loader": "^3.4.2",
     "cypress": "^3.6.1",
     "style-loader": "^1.1.3",
-    "svg-loader": "0.0.2",
     "url-loader": "^3.0.0",
     "webpack": "^4.41.6",
     "webpack-cli": "^3.3.11"
